Update to treeherder-client 3.0.0
Notable changes:
* Performance benefit from connection pooling, if the same
`TreeherderClient` instance is re-used for multiple requests.
* The separate `host` and `protocol` arguments have been combined into a
single `server_url` argument.

See here for more details:
https://bugzilla.mozilla.org/show_bug.cgi?id=1286240

Prior to deploying the source, you'll need to update the copy of
treeherder-client in the virtualenv (it's backwards compatible).
Ideally this would be controlled by a requirements.txt file that was
used as part of an automated deploy.